requirements.txt
streamlit
google-generativeai
graphviz
rich
netmiko